pyproject.toml
[project]
name = "pdf-ocr-rl"
version = "0.1.0"
description = "GRPO fine-tuning of multimodal VLM for PDF-to-markdown conversion"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0",
    "transformers>=4.45",
    "datasets>=2.18",
    "pillow>=10.0",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "requests>=2.31",
    "tqdm>=4.65",
    "Levenshtein>=0.25",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1",
]
train = [
    "unsloth",
    "trl>=0.12",
    "peft>=0.13",
    "bitsandbytes>=0.43",
    "vllm>=0.6",
    "wandb>=0.15",
]
data = [
    "weasyprint>=60.0",
    "markdown>=3.5",
    "pymupdf>=1.24",
    "pdf2image>=1.17",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pdf_ocr_rl"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 120
target-version = "py310"
